feat.添加新的 DataBlock

+++ src/PLCHost/Models/Dtos/PlcInfo.cs
@@ -4,6 +4,6 @@ public record PlcInfo
 {
     public string Ip { get; set; }
     public string CpuType { get; set; }
-    public int Rack { get; set; }
-    public int Slot { get; set; }
+    public short Rack { get; set; }
+    public short Slot { get; set; }
 }
